requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.1.1
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.19
